requirements.txt
azure-containerregistry==1.0.0b7
azure-identity
